Regression test for returning `next()`
Related to #31

NOTE:
Tests are written in sub optimal style where
enforce is last middleware of test app after
mocked "app logic" middlware which is calling next.
This means tests were not testing if `next()` is called
correctly.

To solve this issue for now new extra test for covering this case was
added to prevent further regressions.